fix unused errors in internal/database
