+++ internal/database/BUILD.bazel
@@ -80,8 +80,6 @@ go_library(
         "survey_responses.go",
         "teams.go",
         "temporary_settings.go",
-        "test_util.go",
-        "testing.go",
         "user_credentials.go",
         "user_emails.go",
         "user_roles.go",
@@ -257,6 +255,7 @@ go_test(
         "user_roles_test.go",
         "users_builtin_auth_test.go",
         "users_test.go",
+        "util_test.go",
         "webhook_logs_test.go",
         "webhooks_test.go",
         "zoekt_repos_test.go",
