+++ internal/database/util_test.go
@@ -1,9 +1,29 @@
 package database
 
 import (
+	"context"
 	"encoding/json"
 	"testing"
+
+	"github.com/sourcegraph/sourcegraph/internal/encryption"
 )
+
+func strptr(s string) *string {
+	return &s
+}
+
+func boolptr(b bool) *bool {
+	return &b
+}
+
+func testEncryptionKeyID(key encryption.Key) string {
+	v, err := key.Version(context.Background())
+	if err != nil {
+		panic("why are you sending me a key with an exploding version??")
+	}
+
+	return v.JSON()
+}
 
 func assertJSONEqual(t *testing.T, want, got any) {
 	wantJ := asJSON(t, want)
